Fix circular import issues.
